feat: rename project to cafi
