+++ README.md
@@ -1,18 +1,18 @@
-# Known Acronyms - knAcr
+# Collection acronyms for identification - cafi
 
-[![release: 0.7.0](https://img.shields.io/badge/rel-0.7.0-blue.svg?style=flat-square)](https://github.com/LeibnizDSMZ/knAcr)
+[![release: 0.7.0](https://img.shields.io/badge/rel-0.7.0-blue.svg?style=flat-square)](https://github.com/LeibnizDSMZ/cafi)
 [![MIT LICENSE](https://img.shields.io/badge/License-MIT-brightgreen.svg?style=flat-square)](https://choosealicense.com/licenses/mit/)
 [![DATA LICENSE - CC BY 4.0](https://img.shields.io/badge/Data%20License-CC%20BY%204.0-brightgreen.svg?style=flat-square)](http://creativecommons.org/licenses/by/4.0/)
-[![Documentation Status](https://img.shields.io/badge/docs-GitHub-blue.svg?style=flat-square)](https://LeibnizDSMZ.github.io/knAcr/)
+[![Documentation Status](https://img.shields.io/badge/docs-GitHub-blue.svg?style=flat-square)](https://LeibnizDSMZ.github.io/cafi/)
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg?style=flat-square)](https://github.com/psf/black)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json&style=flat-square)](https://github.com/astral-sh/ruff)
 
-[![main](https://github.com/LeibnizDSMZ/knAcr/actions/workflows/main.yml/badge.svg?branch=main)](https://github.com/LeibnizDSMZ/knAcr/actions/workflows/main.yml)
+[![main](https://github.com/LeibnizDSMZ/cafi/actions/workflows/main.yml/badge.svg?branch=main)](https://github.com/LeibnizDSMZ/cafi/actions/workflows/main.yml)
 
 ---
 
-This project serves as a registry for acronyms used by biological resource centers (BRCs).
+This project serves as a registry for acronyms used by culture collections.
 
 ## Requirements (development)
 
@@ -39,4 +39,4 @@ This project serves as a registry for acronyms used by biological resource cente
 
 All source code is licensed under the MIT License (see LICENSE).
 
-The acronym data inside the data folder (src/knacr/data) is licensed under the [Creative Commons Attribution 4.0 International License](http://creativecommons.org/licenses/by/4.0/) (see LICENSE-CC-BY).
+The acronym data inside the data folder (src/cafi/data) is licensed under the [Creative Commons Attribution 4.0 International License](http://creativecommons.org/licenses/by/4.0/) (see LICENSE-CC-BY).
