+++ configs/dev/mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: Known Acronyms - knAcr
+site_name: Collection acronyms for identification - cafi
 theme:
   name: material
 plugins:
