+++ src/cafi/container/fun/format.py
