+++ src/cafi/errors/custom_exceptions.py
