--- src/knacr/main.py
@@ -1,14 +1,0 @@
-from knacr.errors.custom_exceptions import ValJsonEx
-from knacr.library.loader import load_acr_db
-
-
-def run() -> None:
-    # TODO no real functionality yet
-    try:
-        load_acr_db()
-    except ValJsonEx:
-        print("installed knacr version is not compatible with the latest data")
-
-
-if __name__ == "__main__":
-    run()
